Major refactor
 - Cleaned up some vote code and renamed file in prep for a more generic VoteHandler interface / class.
 - Command files are command implementations. Supporting methods that can be static e.g. non-class-bound are moved to a separate class.
