+++ src/main/java/bot/LookupCommands.java
@@ -3,9 +3,7 @@ package bot;
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.net.URL;
-import java.util.ArrayList;
 import java.util.LinkedList;
-import java.util.regex.*;
 
 import javax.imageio.ImageIO;
 
@@ -30,7 +28,9 @@ implements CommandExecutor{
 		System.out.println("Lookups instantiated.");
 	}
 
-	final String[] dataTypes = {
+	// These datatypes are automatically checked for ttalization and
+	// quotation errors during lookups.
+	static final String[] dataTypes = {
 		"ship",
 		"outfit",
 		"mission",
@@ -88,7 +88,7 @@ implements CommandExecutor{
 		// Check each input for a numeric portion until a non-numeric arg is found.
 		if(args.length > 0)
 			for(String str : args){
-				String number = GetNumeric(str);
+				String number = Helper.GetNumeric(str);
 				if(number.length() > 0)
 					output.append(base + "/" + number + "\n");
 				else
@@ -110,7 +110,7 @@ implements CommandExecutor{
 		// Check each input for a numeric portion until a non-numeric arg is found.
 		if(args.length > 0)
 			for(String str : args){
-				String number = GetNumeric(str);
+				String number = Helper.GetNumeric(str);
 				if(number.length() > 0)
 					output.append(base + "/" + number + "\n");
 				else
@@ -132,7 +132,7 @@ implements CommandExecutor{
 		// Check each input for a hexadecimal hash until a non-hash arg is found.
 		if(args.length > 0)
 			for(String str : args){
-				String hash = GetHash(str);
+				String hash = Helper.GetHash(str);
 				if(hash.length() > 6)
 					output.append(base + hash + "\n");
 				else
@@ -160,15 +160,15 @@ implements CommandExecutor{
 			if(PrintImage(guild, channel, lookupData(request))
 					|| PrintImage(guild, channel, lookupData(variantParsedRequest))
 					|| (IsShipVariantRequest(request)
-								&& PrintImage(guild, channel, lookupData(GetBaseModelName(request))))){
+								&& PrintImage(guild, channel, lookupData(Helper.GetBaseModelName(request))))){
 				printedImage = true;
 			}
 			else
 				message = "There is no image associated with '" + request + "'";
 
 			String output = lookupData(variantParsedRequest);
-			if(!ShouldPrintThis(GetDataType(output)) && output.length() > 0){
-				OutputHelper(channel, "Try '-showdata' for that information.");
+			if(!Helper.ShouldPrintThis(Helper.GetDataType(output)) && output.length() > 0){
+				Helper.OutputHelper(channel, "Try '-showdata' for that information.");
 				return;
 			}
 			if(output.contains("\tdescription")){
@@ -186,7 +186,7 @@ implements CommandExecutor{
 				message = "I could not find anything associated with '" + variantParsedRequest + "'.";
 
 			if(message.length() > 0)
-				OutputHelper(channel, message);
+				Helper.OutputHelper(channel, message);
 		}
 	}
 
@@ -205,7 +205,7 @@ implements CommandExecutor{
 			if(PrintImage(guild, channel, lookupData(request))
 					|| PrintImage(guild, channel, lookupData(variantParsedRequest))
 					|| (IsShipVariantRequest(request)
-								&& PrintImage(guild, channel, lookupData(GetBaseModelName(request))))){
+								&& PrintImage(guild, channel, lookupData(Helper.GetBaseModelName(request))))){
 				printedImage = true;
 			}
 			else
@@ -221,7 +221,7 @@ implements CommandExecutor{
 			else if(!printedImage)
 				message += ", but I did find this:\n\n";
 
-			OutputHelper(channel, message + output);
+			Helper.OutputHelper(channel, message + output);
 		}
 	}
 
@@ -237,11 +237,11 @@ implements CommandExecutor{
 			if(PrintImage(guild, channel, lookupData(request))
 					|| PrintImage(guild, channel, lookupData(ParseVariants(request)))
 					|| (IsShipVariantRequest(request)
-								&& PrintImage(guild, channel, lookupData(GetBaseModelName(request))))){
+								&& PrintImage(guild, channel, lookupData(Helper.GetBaseModelName(request))))){
 				// This request was handled.
 			}
 			else
-				OutputHelper(channel, "I could not find an image associated with '" + request + "'.");
+				Helper.OutputHelper(channel, "I could not find an image associated with '" + request + "'.");
 		}
 	}
 
@@ -259,7 +259,7 @@ implements CommandExecutor{
 			if(output.length() < 1){
 				output = "I could not find any data associated with '" + request + "'.";
 			}
-			OutputHelper(channel, output);
+			Helper.OutputHelper(channel, output);
 		}
 	}
 
@@ -290,7 +290,7 @@ implements CommandExecutor{
 		// Remove any leading or trailing spaces.
 		lookup = lookup.trim();
 		// The first word of the lookup may be a supported dataType.
-		int countWords = 1 + CountOf(lookup, ' ');
+		int countWords = 1 + Helper.CountOf(lookup, ' ');
 		String category = "";
 		if(countWords > 1){
 			category = lookup.substring(0, lookup.indexOf(" ")).toLowerCase();
@@ -356,7 +356,7 @@ implements CommandExecutor{
 		}
 		// The input may not have been capitalized correctly.
 		if(helper){
-			lookup = CapitalizeWords(lookup);
+			lookup = Helper.CapitalizeWords(lookup);
 			return checkLookup(dataType, lookup, false);
 		}
 
@@ -365,30 +365,13 @@ implements CommandExecutor{
 
 
 
-	// Send the message 'output' to the desired channel, cutting into
-	// multiple messages as needed.
-	private static void OutputHelper(MessageChannel channel, String output){
-		if(output.length() < 1993){
-			channel.sendMessage(":\n```" + output + "```").queue();
-		}
-		else{
-			int cut = output.lastIndexOf('\n', 0 + 1992);
-			String o = output.substring(0, cut);
-			channel.sendMessage(":\n```" + o + "```").queue(x -> {
-				OutputHelper(channel, output.substring(cut + 1));
-			});
-		}
-	}
-
-
-
 	// Check the string for image characteristics, and if found, print the image
 	// to the specified channel & return true. Returns false for no image or no
 	// valid image ending.
 	private boolean PrintImage(Guild guild, MessageChannel channel, String input){
-		if(HasImageToPrint(input)){
-			String imageName = GetImageName(input);
-			String filepath = urlEncode(CONTENT_URL + "/images/" + imageName);
+		if(Helper.HasImageToPrint(input)){
+			String imageName = Helper.GetImageName(input);
+			String filepath = Helper.urlEncode(CONTENT_URL + "/images/" + imageName);
 			String ending = GetImageEnding(filepath);
 			if(ending.length() > 0){
 				EmbedBuilder eb = new EmbedBuilder();
@@ -444,58 +427,6 @@ implements CommandExecutor{
 
 
 
-	public static String urlEncode(String url){
-		return url.replace(" ", "%20");
-	}
-
-
-
-	// Check the string for image indicators. Returns false if there is no image.
-	public static boolean HasImageToPrint(String input){
-		return input.contains("\tsprite ") || input.contains("\tthumbnail ")
-				|| input.contains("\tlandscape ") || input.contains("\tscene ");
-	}
-
-
-
-	// Check the string for a space or dash character and if present,
-	// capitalize the next letter. Returns the string with captialized words.
-	private static String CapitalizeWords(String input){
-		int countWords = 1 + CountOf(input, ' ') + CountOf(input, '-');
-		char[] ic = input.toCharArray();
-		ic[ic[0] == '"' ? 1 : 0] = Character.toUpperCase(ic[ic[0] == '"' ? 1 : 0]);
-		boolean hasDash = input.indexOf("-") > -1;
-		boolean hasSpace = input.indexOf(" ") > -1;
-		if(--countWords > 0){
-			int index = input.indexOf(" ");
-			if(hasDash)
-				index = Math.min(index, input.indexOf("-"));
-			for(int i = 0; i < countWords; ++i){
-				++index;
-				if(ic[index] == '(' || ic[index] == ')' || ic[index] == '"' || ic[index] == '-')
-					++index;
-				ic[index] = Character.toUpperCase(ic[index]);
-				// Find next word break.
-				int s = input.indexOf(" ", index);
-				int d = input.indexOf("-", index);
-				hasDash &= d > -1;
-				hasSpace &= s > -1;
-				if(hasDash && hasSpace)
-					index = (s < d) ? s : d;
-				else if(hasDash)
-					index = d;
-				else if(hasSpace)
-					index = s;
-				else
-					break;
-			}
-		}
-
-		return new String(ic);
-	}
-
-
-
 	// Check if the request is for a ship variant.
 	private boolean IsShipVariantRequest(String request){
 		// All Endless Sky ship variants have the name in parentheses by convention.
@@ -506,7 +437,7 @@ implements CommandExecutor{
 		// not a ship variant, or is already a ship variant with proper formatting.
 		// e.g. '"base name" "base name (variant)"', 'Thruster (Stellar Class)', or
 		// '"Thruster (Stellar Class)"'
-		String base = GetBaseModelName(request);
+		String base = Helper.GetBaseModelName(request);
 		if(lookupData(request).length() > 0)
 			return request.indexOf(base) != request.lastIndexOf(base);
 
@@ -526,7 +457,7 @@ implements CommandExecutor{
 	// formatted, returns the request in the proper ship variant format.
 	private String ParseVariants(String input){
 		if((input.indexOf('(') > 0 || input.indexOf(')') > 0) && lookupData(input).length() < 1){
-			String baseModel = GetBaseModelName(input);
+			String baseModel = Helper.GetBaseModelName(input);
 			if(baseModel.length() > 0
 					&& input.indexOf(baseModel) == input.lastIndexOf(baseModel)){
 				input = "\"" + baseModel.replace("\"", "") + "\" \"" + input.replace("\"", "") + "\"";
@@ -538,108 +469,6 @@ implements CommandExecutor{
 
 
 
-	// Count the number of the given character in the given string.
-	public static int CountOf(String input, char token){
-		int count = 0;
-		for(char c : input.toCharArray()){
-			if(c == token)
-				++count;
-		}
-
-		return count;
-	}
-
-
-
-	// Returns the bare image name without quotes, or a nullstring if no image.
-	public static String GetImageName(String text){
-		int start = 0;
-		if(text.contains("\tthumbnail"))
-			start = text.indexOf("thumbnail") + 10;
-		else if(text.contains("\tsprite"))
-			start = text.indexOf("sprite") + 7;
-		else if(text.contains("\tlandscape"))
-			start = text.indexOf("landscape") + 10;
-		else if(text.contains("\tscene"))
-			start = text.indexOf("scene") + 6;
-		else
-			return "";
-
-		int end = text.indexOf('\n', start);
-
-		return text.substring(start, end).replace("\"", "");
-	}
-
-
-
-	// Called for ship variants in order to obtain the base model name. Returns
-	// an unquoted ship, e.g. "Falcon (Plasma)" -> Falcon, or "Marauder Falcon
-	// (Engines)" -> Marauder Falcon. Will not work for variants which do not use
-	// the variant name convention of "Base Ship Name (varied text)"
-	// Returns nullstring if no match.
-	public static String GetBaseModelName(String text){
-		int end = text.indexOf('(');
-		if(end < 0)
-			return "";
-
-		return text.substring(0, end-1).replaceAll("\"", "").trim();
-	}
-
-
-
-	// Returns the bit that comes before the searched request string.
-	// e.g. "mission", "ship", "fleet", "outfit"
-	public static String GetDataType(String output){
-		if(output.length() < 1 || output.indexOf(" ") < 1)
-			return "";
-		int start = (output.indexOf('\n') == 0 || output.indexOf('\t') == 0) ? 1 : 0;
-		return output.substring(start, output.indexOf(" "));
-	}
-
-
-
-	// Things that don't generally have images or descriptions probably shouldn't
-	// get printed from -lookup.
-	private static boolean ShouldPrintThis(String lookupType){
-		if(lookupType.length() < 1)
-			return false;
-
-		switch(lookupType.toLowerCase()){
-			case "mission":
-				return false;
-			case "event":
-				return false;
-			case "fleet":
-				return false;
-			default:
-				return true;
-		}
-	}
-
-
-
-	// Return the numeric portion of the passed argument
-	private static String GetNumeric(String arg){
-		arg = arg.replaceAll("\"", "").replaceAll("\'", "").trim();
-		Pattern regex = Pattern.compile("\\d{1,}");
-		Matcher result = regex.matcher(arg);
-		String numberPart = result.find() ? result.group() : "";
-		return numberPart;
-	}
-	
-	
-	
-	// Return the hexadecimal portion of the passed argument.
-	private static String GetHash(String arg){
-		arg = arg.replaceAll("\"", "").replaceAll("\'", "").trim();
-		Pattern regex = Pattern.compile("[\\dA-Fa-f]+");
-		Matcher result = regex.matcher(arg);
-		String hash = result.find() ? result.group() : "";
-		return hash.length() > 6 ? hash : "";
-	}
-
-
-
 	// Generate a quote from the named person, using their built-in phrases.
 	// Phrases can be nested!! TODO: Move choice instantiation into readData.
 	public String generateQuote(String person){
@@ -647,8 +476,8 @@ implements CommandExecutor{
 		boolean hasPhrase = personData.indexOf("phrase\n") > -1 || personData.indexOf("phrase ") > -1;
 		if(!hasPhrase)
 			return "";
-		int tabDepth = GetIndentLevel(personData, "phrase");
-		ArrayList<String> strKeys = new ArrayList<String>();
+		int tabDepth = Helper.GetIndentLevel(personData, "phrase");
+		LinkedList<String> strKeys = new LinkedList<String>();
 
 		// Outermost 'phrase'
 		StringBuilder key = new StringBuilder();
@@ -682,54 +511,5 @@ implements CommandExecutor{
 		// and then passing the substring to MakeChoices
 
 		return Integer.toString(tabDepth);
-	}
-
-
-
-	// Call this for each word in a phrase, giving it the bits
-	// between 'word', 'phrase', or the next data entry.
-	private static String[] MakeChoices(String wordlist){
-		final int MAXPOS = 10000000;
-		int start = wordlist.indexOf("\tword\n");
-		ArrayList<String> choices = new ArrayList<String>();
-		boolean reachedEnd = false;
-		while(!reachedEnd){
-			int startQuote = wordlist.indexOf("\t\"", start);
-			int startTick = wordlist.indexOf("\t`", start);
-			int startWord = Integer.min(startQuote > -1 ? startQuote : MAXPOS, startTick > -1 ? startTick : MAXPOS);
-			if(startWord == MAXPOS){
-				reachedEnd = true;
-				break;
-			}
-			boolean findTick = startQuote == -1 || (startTick > -1 && startTick < startQuote);
-			int endWord = wordlist.indexOf(findTick ? "`\n" : "\"\n", findTick ? ++startTick : ++startQuote);
-			choices.add(wordlist.substring(findTick ? startTick : startQuote, ++endWord));
-			start = endWord;
-		}
-		if(!choices.isEmpty())
-			return choices.toArray(new String[choices.size()]);
-		else
-			return new String[0];
-	}
-
-
-
-	// Return the number of tabs that preceed the specified indented word in the
-	// datastring. Phrases are generally 0 or 1 level, while word demarcations
-	// are generally 1 or 2 indent levels.
-	private static int GetIndentLevel(String data, String word){
-		int level = -1;
-		int start = data.indexOf(word);
-		if(start < 0)
-			return level;
-
-		int lineStart = data.lastIndexOf("\n", start);
-		// If the data string starts with this word, return 0, even if there is no
-		// newline preceeding it.
-		if(lineStart < 0 || start == 0)
-			return 0;
-
-		String line = data.substring(lineStart, start + word.length() + 1);
-		return CountOf(line,'\t');
 	}
 }
